pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "corax"
authors = [{ name = "Yicheng Luo", email = "ethanluoyc@gmail.com" }]
requires-python = ">=3.9"
dependencies = ["absl-py", "dm-env", "numpy", "dm-tree"]
dynamic = ["version"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: Implementation :: CPython",
]

[tool.hatch.version]
source = "vcs"

[project.optional-dependencies]
tf = [
  # TF dependencies
  "tensorflow-cpu~=2.13.0",
  "tensorflow-probability~=0.21.0",
  "tensorflow-datasets~=4.9.3",
  "dm-reverb~=0.12.0",
  "rlds",
]
jax = [
  "corax[tf]",
  "jax",
  "dm-haiku",
  "flax",
  "optax",
  "rlax",
  "chex",
  "dm_env_wrappers",
]
all = ["corax[tf,jax]"]
test = [
  "corax[all]",
  "pytest",
  "dill",           # required for tfds tests
  "pytest-xdist",
  "wandb",
  "ml-collections",
  "scipy>=1.6.0",
  # Environments
  "gym>=0.21.0,<0.24.0",
  "gymnasium",
  "ott-jax",
  "dm-control",
]
dev = ["corax[test]", "black", "ruff"]

[tool.hatch.build]
directory = "dist"

[tool.hatch.build.targets.wheel]
exclude = ["dummy_data/", "*_test.py"]

[tool.black]
line-length = 88
target-version = ['py39', 'py310']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.github
  | \.venv
  | docs
  | third_party
  | scratch
)/
'''

[tool.ruff]
line-length = 120
select = [
  "E", # pycodestyle
  "F", # pyflakes
  "I",
]
ignore = ["E731"]
exclude = ["third_party", ".venv"]

[tool.ruff.extend-per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["corax"]
force-single-line = true
single-line-exclusions = ["typing"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.pyright]
include = ["corax"]
exclude = ["**/node_modules", "**/__pycache__"]
ignore = ["corax/adders", "corax/datasets/tfds"]

reportMissingImports = true
reportMissingTypeStubs = false

pythonVersion = "3.9"
pythonPlatform = "Linux"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["corax"]
